+++ cloud/azure/runtime/gateway/http.go
@@ -29,8 +29,8 @@ import (
 
 	"github.com/nitrictech/nitric/cloud/azure/runtime/resource"
 	base_http "github.com/nitrictech/nitric/cloud/common/runtime/gateway"
-	"github.com/nitrictech/nitric/cloud/common/runtime/logger"
 	"github.com/nitrictech/nitric/core/pkg/gateway"
+	"github.com/nitrictech/nitric/core/pkg/logger"
 	schedulespb "github.com/nitrictech/nitric/core/pkg/proto/schedules/v1"
 	storagepb "github.com/nitrictech/nitric/core/pkg/proto/storage/v1"
 	topicpb "github.com/nitrictech/nitric/core/pkg/proto/topics/v1"
